persistent keepalive: use unsigned long to avoid multiplication in hotpath

+++ src/peer.h
@@ -29,7 +29,7 @@ struct wireguard_peer {
 	uint64_t rx_bytes, tx_bytes;
 	struct timer_list timer_retransmit_handshake, timer_send_keepalive, timer_new_handshake, timer_kill_ephemerals, timer_persistent_keepalive;
 	unsigned int timer_handshake_attempts;
-	uint16_t persistent_keepalive_interval;
+	unsigned long persistent_keepalive_interval;
 	bool timer_need_another_keepalive;
 	struct timeval walltime_last_handshake;
 	struct sk_buff_head tx_packet_queue;
